refactor(core): move security checks to core
