+++ package.json
@@ -61,7 +61,6 @@
     "@simplewebauthn/typescript-types": "^8.3.4",
     "@sunknudsen/totp": "^1.1.0",
     "@tsconfig/node22": "^22.0.0",
-    "@types/bcryptjs": "^2.4.6",
     "@types/console-log-level": "^1.4.5",
     "@types/cookie-parser": "^1.4.7",
     "@types/ejs": "^3.1.5",
@@ -70,7 +69,6 @@
     "@types/morgan": "^1.9.9",
     "await-to-js": "^3.0.0",
     "axios": "^1.7.7",
-    "bcryptjs": "^2.4.3",
     "body-parser": "^1.20.3",
     "connect-redis": "^7.1.1",
     "console-log-level": "^1.4.1",
@@ -85,13 +83,11 @@
     "html-to-text": "^9.0.5",
     "http-errors": "^2.0.0",
     "ioredis": "^5.4.1",
-    "is-disposable-email-domain": "^1.0.7",
     "lodash": "^4.17.21",
     "lodash-es": "^4.17.21",
     "moment": "^2.30.1",
     "moment-timezone": "^0.5.45",
     "morgan": "^1.10.0",
-    "nanoid": "^3.3.6",
     "nocache": "^4.0.0",
     "node-pg-migrate": "^7.6.1",
     "nodemailer": "^6.9.15",
@@ -112,28 +108,28 @@
     "@changesets/cli": "^2.27.10",
     "@simplewebauthn/types": "^10.0.0",
     "@sinonjs/fake-timers": "^11.2.2",
-    "@types/chai": "^5.0.0",
+    "@types/chai": "^5.0.1",
     "@types/chai-as-promised": "^7.1.8",
     "@types/html-to-text": "^9.0.4",
     "@types/http-errors": "^2.0.4",
     "@types/lodash": "^4.17.10",
     "@types/lodash-es": "^4.17.12",
-    "@types/mocha": "^10.0.7",
-    "@types/node": "^22.1.0",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^22.10.2",
     "@types/nodemailer": "^6.4.16",
     "@types/oidc-provider": "^8.5.2",
     "@types/qrcode": "^1.5.5",
     "@types/sinonjs__fake-timers": "^8.1.5",
     "axe-core": "^4.8.4",
-    "chai": "^5.1.1",
-    "chai-as-promised": "^8.0.0",
+    "chai": "^5.1.2",
+    "chai-as-promised": "^8.0.1",
     "concurrently": "^9.0.1",
     "copy-and-watch": "^0.1.6",
     "csv": "^6.3.9",
     "cypress": "^13.17.0",
     "cypress-axe": "^1.5.0",
     "cypress-maildev": "^1.3.2",
-    "mocha": "^10.7.3",
+    "mocha": "^11.0.1",
     "nock": "^13.5.4",
     "prettier": "^3.4.2",
     "prettier-plugin-organize-imports": "^4.1.0"
