+++ packages/core/package.json
@@ -46,13 +46,21 @@
   },
   "dependencies": {
     "@zootools/email-spell-checker": "^1.12.0",
-    "is-disposable-email-domain": "^1.0.7"
+    "bcryptjs": "^2.4.3",
+    "is-disposable-email-domain": "^1.0.7",
+    "lodash-es": "^4.17.21",
+    "nanoid": "^5.0.9",
+    "tld-extract": "^2.1.0"
   },
   "devDependencies": {
     "@tsconfig/node22": "^22.0.0",
-    "@types/mocha": "^10.0.7",
-    "chai": "^5.1.1",
-    "mocha": "^10.7.3",
+    "@types/bcrypt": "^5.0.2",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/lodash-es": "^4.17.12",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^22.10.2",
+    "chai": "^5.1.2",
+    "mocha": "^11.0.1",
     "tsx": "^4.19.2"
   },
   "publishConfig": {
