+++ packages/core/src/data/diceware-wordlist-fr-alt.ts
