+++ packages/core/src/security/generate-token.ts
@@ -1,0 +1,9 @@
+//
+
+import { nanoid } from "nanoid";
+
+//
+
+export function generateToken() {
+  return nanoid(64);
+}
