+++ packages/core/src/security/is-name-valid.ts
@@ -1,0 +1,5 @@
+//
+
+export function isNameValid(name: string) {
+  return !!name.match(/^[^$&+:;=?@#|<>.^*()%!\d_\[\]{}\\\/"`~]*$/);
+}
