+++ packages/core/src/security/is-visible-string.ts
@@ -1,0 +1,7 @@
+//
+
+const VISIBLE_CHAR_REGEX = /[^\s\p{Cf}\p{Cc}\p{Zl}\p{Zp}]/u;
+
+export function isVisibleString(input: string) {
+  return VISIBLE_CHAR_REGEX.test(input);
+}
