+++ packages/core/src/security/owasp-password-strength-tester.ts
