+++ packages/core/src/security/validate-password.ts
@@ -1,0 +1,16 @@
+//
+
+import bcrypt from "bcryptjs";
+
+//
+
+export function validatePassword(
+  plainPassword: string,
+  storedHash: string | null,
+) {
+  if (!plainPassword || !storedHash) {
+    return false;
+  }
+
+  return bcrypt.compare(plainPassword || "", storedHash);
+}
