+++ packages/core/tsconfig.lib.json
@@ -1,16 +1,9 @@
 {
   "compilerOptions": {
     "outDir": "./dist",
-    "rootDir": "./src",
-    "types": [
-      "./types"
-    ]
+    "rootDir": "./src"
   },
-  "exclude": [
-    "src/**/*.test.ts"
-  ],
+  "exclude": ["src/**/*.test.ts"],
   "extends": "./tsconfig.json",
-  "include": [
-    "src"
-  ]
+  "include": ["src"]
 }
