+++ packages/core/types/is-disposable-email-domain.d.ts
@@ -1,0 +1,5 @@
+//
+
+declare module "is-disposable-email-domain" {
+  function isFree(email: string): boolean;
+}
