+++ src/connectors/api-annuaire-service-public.ts
@@ -1,3 +1,4 @@
+import { isEmailValid } from "@gouvfr-lasuite/proconnect.core/security";
 import axios, { AxiosError, type AxiosResponse } from "axios";
 import { isEmpty, isString } from "lodash-es";
 import {
@@ -12,7 +13,6 @@ import {
   ApiAnnuaireTooManyResultsError,
 } from "../config/errors";
 import { logger } from "../services/log";
-import { isEmailValid } from "../services/security";
 
 // more info at https://api-lannuaire.service-public.fr/api/explore/v2.1/console
