+++ src/managers/organization/join.ts
@@ -1,3 +1,4 @@
+import { isEmailValid } from "@gouvfr-lasuite/proconnect.core/security";
 import { Welcome } from "@gouvfr-lasuite/proconnect.email";
 import * as Sentry from "@sentry/node";
 import { isEmpty, some } from "lodash-es";
@@ -53,7 +54,6 @@ import {
   isEtablissementScolaireDuPremierEtSecondDegre,
   isSmallAssociation,
 } from "../../services/organization";
-import { isEmailValid } from "../../services/security";
 import { unableToAutoJoinOrganizationMd } from "../../views/mails/unable-to-auto-join-organization";
 import { getOrganizationsByUserId, markDomainAsVerified } from "./main";
