+++ src/managers/organization/official-contact-email-verification.ts
@@ -1,3 +1,4 @@
+import { generateDicewarePassword } from "@gouvfr-lasuite/proconnect.core/security";
 import { OfficialContactEmailVerification } from "@gouvfr-lasuite/proconnect.email";
 import { isEmpty } from "lodash-es";
 import { HOST } from "../../config/env";
@@ -20,7 +21,6 @@ import {
   isCommune,
   isEtablissementScolaireDuPremierEtSecondDegre,
 } from "../../services/organization";
-import { generateDicewarePassword } from "../../services/security";
 
 const OFFICIAL_CONTACT_EMAIL_VERIFICATION_TOKEN_EXPIRATION_DURATION_IN_MINUTES = 60;
 
@@ -94,8 +94,7 @@ export const sendOfficialContactEmailVerificationEmail = async ({
     return { codeSent: false, contactEmail, libelle };
   }
 
-  const official_contact_email_verification_token =
-    await generateDicewarePassword();
+  const official_contact_email_verification_token = generateDicewarePassword();
 
   await updateUserOrganizationLink(organization_id, user_id, {
     official_contact_email_verification_token,
